server/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
# AI & LangChain
langchain<0.3.0
langchain-community<0.3.0
langchain-ollama
langchain-google-genai
langchain-huggingface
langchain-openai
# Processing
pypdf
faiss-cpu
python-dotenv
gTTS
sentence-transformers
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
psycopg2-binary
email-validator
